requirements.txt
Django
python-dotenv
google-generativeai